% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countEdges.R
\name{countEdges}
\alias{countEdges}
\title{Count the edges information of an initial
random tensor graph.}
\usage{
countEdges(X, G)
}
\arguments{
\item{X}{A list of matrices}

\item{G}{The core tensor}
}
\value{
a 2-dim vector describing average degree
}
\description{
Count the edges information of an initial
random tensor graph.
}
