% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRTG.R
\name{fastRTG}
\alias{fastRTG}
\title{Title}
\usage{
fastRTG(X, G, avgDeg = NULL, PoissonEdges = TRUE, returnParameters = FALSE)
}
\arguments{
\item{X}{a list of matrices X_i's. Each X_i is a \eqn{n_i by k_i} matrix.
n_i is the dimension size in the i-th mode. k_i is the number of latent
factors in the i-th mode. If X only has one matrix, the default reads it
as a 3-mode tensor with same X_1 in all three modes.}

\item{G}{the core tensor. Should not contain negative values.}

\item{avgDeg}{specifies the expected degree.}

\item{PoissonEdges}{boolean indicator. Allow poisson edges if TRUE, otherwise only binary edges.}

\item{returnParameters}{return parameter list or not.}
}
\value{

}
\description{
Title
}
