% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbmt.R
\name{sbmt}
\alias{sbmt}
\title{Sampling Stochstic Tensor Block Models.}
\usage{
sbmt(
  n,
  Pi,
  G,
  PoissonEdges = FALSE,
  avgDeg = NULL,
  returnParameters = FALSE,
  parametersOnly = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{

}
