% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRTG.R
\name{fastRTG}
\alias{fastRTG}
\title{Title}
\usage{
fastRTG(X, G, avgDeg = NULL, PoissonEdges = TRUE, returnParameters = FALSE)
}
\arguments{
\item{X}{a list of matrices X_i's. Each X_i is a \eqn{n_i by k_i} matrix.
n_i is the dimension size in the i-th mode. k_i is the number of latent
factors in the i-th mode. If X only has one matrix, the default reads it
as a 3-mode tensor with same X_1 in all three modes.}

\item{G}{the core tensor. Should not contain negative values.}

\item{avgDeg}{specifies the expected degree.}

\item{PoissonEdges}{boolean indicator. Allow poisson edges if TRUE, otherwise only binary edges.}

\item{returnParameters}{return parameter list or not.}
}
\value{
if returnParameters is TRUE, returns a list containing sampled tensor
as well as the ground truth latent factors X, core tensor G.
}
\description{
Title
}
\examples{
G = as.tensor(array(rgamma(120,1,1), dim = c(2,3,4,5)))
X[[1]] = matrix(abs(rnorm(200)),100,2)
X[[2]] = matrix(rgamma(300,1,1),100,3)
X[[3]] = matrix(rf(400,3,4),100,4)
X[[4]] = matrix(1,100, 5)
sampleTensor <- fastRTG(X, G, avgDeg = 1000, returnParameters = TRUE)
}
