% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRTG.R
\name{fastRTG}
\alias{fastRTG}
\title{Fast Sampling for Random Tensor Graphs}
\usage{
fastRTG(X, G, avgDeg = NULL, PoissonEdges = TRUE, returnParameters = FALSE)
}
\arguments{
\item{X}{a list of matrices X_i's. Each X_i is a \eqn{n_i by k_i} matrix.
n_i is the dimension size in the i-th mode. k_i is the number of latent
factors in the i-th mode. If X only has one matrix, the default reads it
as a 3-mode tensor with same X_1 in all three modes.}

\item{G}{the core tensor. Should not contain negative values.}

\item{avgDeg}{specifies the expected degree.}

\item{PoissonEdges}{boolean indicator. Allow poisson edges if TRUE, otherwise only binary edges.}

\item{returnParameters}{return parameter list or not.}
}
\value{
if returnParameters is TRUE, returns a list containing sampled tensor
as well as the ground truth latent factors X, core tensor G.
}
\description{
Provide efficient algorithm to sample various random tensor graphs. This
function allows general low-rank tensor setup in the sense that the core
tensor can has arbitrary number of modes and different number of latent
factors in each modes (Tucker Decomposition).
}
\examples{
G = as.tensor(array(rgamma(120,1,1), dim = c(2,3,4,5)))
X[[1]] = matrix(abs(rnorm(20)),100,2)
X[[2]] = matrix(rgamma(30,1,1),100,3)
X[[3]] = matrix(rf(40,3,4),100,4)
X[[4]] = matrix(1,10, 5)
sampleTensor <- fastRTG(X, G, avgDeg = 10, returnParameters = TRUE)
}
