% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRTG.R
\name{fastRTG}
\alias{fastRTG}
\title{Fast Sampling for Random Tensor Graphs}
\usage{
fastRTG(X, G, sparsity = NULL, PoissonEdges = TRUE, returnParameters = FALSE)
}
\arguments{
\item{X}{a list of matrices X_i's. Each X_i is a \eqn{n_i by k_i} matrix.
n_i is the dimension size in the i-th mode. k_i is the number of latent
factors in the i-th mode. If X only has one matrix, the default reads it
as a 3-mode tensor with same X_1 in all three modes.}

\item{G}{the core tensor represented by a multi-dimensional array. Should not contain negative values.}

\item{PoissonEdges}{boolean indicator. Allow poisson edges if \code{TRUE}, otherwise only binary edges.}

\item{returnParameters}{return parameter list or not.}

\item{avgDeg}{specifies the expected degree.}
}
\value{
if returnParameters is TRUE, returns a list containing sampled tensor
as well as the ground truth latent factors X, core tensor G. The sampled tensor
is a tensorr::sptensor object. The latent factors is a list of matrices. The core tensor
is a rTensor::Tensor object.
}
\description{
Provide efficient algorithm to sample various random tensor graphs. This
function allows general low-rank tensor setup in the sense that the core
tensor can has arbitrary number of modes and different number of latent
factors in each modes (Tucker Decomposition).
}
\examples{
G = array(rgamma(120,1,1), dim = c(2,3,4,5))
X[[1]] = matrix(abs(rnorm(20)),10,2)
X[[2]] = matrix(rgamma(60,1,1),20,3)
X[[3]] = matrix(rf(120,3,4),30,4)
X[[4]] = matrix(1,40, 5)
sampleTensor <- fastRTG(X, G, sparsity = 0.01, returnParameters = TRUE)
}
