% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbmt.R
\name{sbmt}
\alias{sbmt}
\title{Sampling Stochstic Tensor Block Models.}
\usage{
sbmt(
  n,
  Pi,
  G,
  PoissonEdges = FALSE,
  avgDeg = NULL,
  returnParameters = FALSE,
  parametersOnly = FALSE,
  ...
)
}
\arguments{
\item{n}{a vector of positive integers specifying the dimensions sizes of each mode.}

\item{Pi}{a vector with same size as the number of clusters specifying the cluster weights.}

\item{G}{the core tensor, should be an rTensor tensor type.}

\item{PoissonEdges}{boolean indicator. If TRUE, elements of A allows multiple same edges. IF FALSE, A should be a binary tensor.}

\item{avgDeg}{an integer specifying the expected degree.}

\item{returnParameters}{logical. Return the parameters or not.}

\item{parametersOnly}{logical. Only return the parameters or not.}

\item{...}{other parameters.}
}
\value{
A list of three items: The sampled random tensor. The latent factors Zs. The ground truth core tensor G.
}
\description{
Suppose the binary (Poisson) tensor A has expectation structure:
\deqn{E(A) = [G;Z_1,Z_2,Z_3,...Z_m],}
where each row of Z_i's has exactly one non-zero element with value equals to 1.
}
