% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbmt.R
\name{sbmt}
\alias{sbmt}
\title{Sampling Stochstic Tensor Block Models.}
\usage{
sbmt(
  n,
  Pi,
  G,
  PoissonEdges = FALSE,
  sparsity = NULL,
  returnParameters = FALSE,
  parametersOnly = FALSE,
  ...
)
}
\arguments{
\item{n}{a vector of positive integers specifying the dimensions sizes of each mode.}

\item{Pi}{a vector with same size as the number of clusters specifying the cluster weights.}

\item{G}{the core tensor in a multi-dimensional array.}

\item{PoissonEdges}{boolean indicator. If TRUE, elements of A allows multiple same edges. IF FALSE, A should be a binary tensor.}

\item{returnParameters}{logical. Return the parameters or not.}

\item{parametersOnly}{logical. Only return the parameters or not.}

\item{...}{other parameters.}

\item{avgDeg}{an integer specifying the expected degree.}
}
\value{
A list of three items: The sampled random tensor. The latent factors Zs. The ground truth core tensor G.
}
\description{
Suppose the binary (Poisson) tensor A has expectation structure:
\deqn{E(A) = [G;Z_1,Z_2,Z_3,...Z_m],}
where each row of Z_i's has exactly one non-zero element with value equals to 1.
}
\examples{
n=10
G = array(abs(rnorm(27)),dim = rep(3,3))
Pi = c(0.3,0.3,0.4)
sampleSBM = sbmt(n, Pi, G, sparsity = 0.01)
}
